Graphic layers, added layers, hunt area query
- Created separate graphic layers for each tool (hunt area query, place
search, coordinate search, coordinate search label, measurements, and
draw).
- Changed PictureMarkerSymbols to SimpleMarkerSymbols.  The PMS does not
render correctly on the Export Web Map PDF.
- Added administrative boundary layers (IDFG regions, WMAs, Couties, BLM
field offices, Tribal lands, Wilderness Areas, National Forest
Boundaries) and the Land Management layer.
- Added code that toggled legend layers on/off by clicking the layer
label name as well as the checkbox.
- Added code that gets variables to run a query to highlight a hunt
area.  The variables are taken from a URL that is associated with a map
the user clicks on in the IDFG Hunt Planner application.
- Changed graphic layer color from red or black to purple or yellow
(easier for the color-blind community to depict).
- When the user now searches for a coordinate, it zooms to it AND adds
the long, lat coordinate as a label
- Added a tool to add text to the map using the draw toolbar point tool.