Proxy page & print PDF hyperlink/ ts url
To combat errors users are receiving while trying to use the create PDF
tool, we are now using our own proxy page.
To keep PDF window from being blocked by popup blockers, we added a
hyperlink to the print modal for users to click to open the PDF in a new
window.
Also, add a timestamp to the printTask response url to hopefully to keep
older version of the PDF from showing up when trying to create a new PDF
map.