Collapsible Legend, Share Map View
The 'Turn Layers On/Off' has been changed to a collapsible window in the
sidebar so users can see the legend while viewing the map on larger
screens.

The 'Share the Map View' tool was created so users can share a specific
map view based on a center coordinate and zoom level.

'Game Distribution' dropdown list was added to the 'Highlight a Hunt
Area' modal.  Users can select game species by name to see their general
distribution in Idaho.